+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# HEAD
+
+- feat: support links and inline code
+
 # v1.0.0 (2019-05-25)
 
 - initial release (the version I used at RubyKaigi2019)